refactor: 💡 move the reference lines logic to another file
